LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-used-values-expected.txt


FAIL Use the last value from view-timeline-axis if omitted assert_equals: expected "-1" but got "0"
FAIL Use the last value from view-timeline-inset if omitted assert_equals: expected "-1" but got "0"
